+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next build on `develop` should refer to version 4.2.0 -->
 ### Fixed
 - Move "getContentString" method to Localization service [#2291](https://github.com/rokwire/illinois-app/issues/2291).
+### Changed
+- The text for logged out messages [#2290](https://github.com/rokwire/illinois-app/issues/2290).
 
 ## [4.1.19] - 2022-08-15
 ### Changed